Restructure simple framework for asserting SIARD archives
