+++ src/integrationTest/java/ch/admin/bar/siard2/cmd/OracleUploadDownloadSiardProjectIT.java
@@ -4,7 +4,7 @@ import ch.admin.bar.siard2.cmd.utils.ConsoleLogConsumer;
 import ch.admin.bar.siard2.cmd.utils.SiardProjectExamples;
 import ch.admin.bar.siard2.cmd.utils.SqlScripts;
 import ch.admin.bar.siard2.cmd.utils.TestResourcesResolver;
-import ch.admin.bar.siard2.cmd.utils.siard.SiardArchiveAssertions;
+import ch.admin.bar.siard2.cmd.utils.siard.assertions.SiardArchiveAssertions;
 import ch.admin.bar.siard2.cmd.utils.siard.SiardArchivesHandler;
 import lombok.val;
 import org.junit.Assert;
