--- src/integrationTest/java/ch/admin/bar/siard2/cmd/utils/siard/model/QualifiedTableId.java
@@ -1,12 +1,0 @@
-package ch.admin.bar.siard2.cmd.utils.siard.model;
-
-import lombok.Builder;
-import lombok.NonNull;
-import lombok.Value;
-
-@Value
-@Builder
-public class QualifiedTableId {
-    @NonNull Id<SchemaMetaData> schemaId;
-    @NonNull Id<TableMetaData> tableId;
-}
