+++ src/integrationTest/java/ch/admin/bar/siard2/cmd/utils/siard/model/SiardArchive.java
@@ -1,6 +1,7 @@
 package ch.admin.bar.siard2.cmd.utils.siard.model;
 
-import ch.admin.bar.siard2.cmd.utils.siard.model.content.SiardContent;
+import ch.admin.bar.siard2.cmd.utils.siard.model.header.Metadata;
+import ch.admin.bar.siard2.cmd.utils.siard.model.content.Content;
 import ch.admin.bar.siard2.cmd.utils.siard.update.Updatable;
 import ch.admin.bar.siard2.cmd.utils.siard.update.Updater;
 import lombok.NonNull;
@@ -9,8 +10,8 @@ import lombok.val;
 
 @Value
 public class SiardArchive implements Updatable<SiardArchive> {
-    @NonNull SiardMetadata siardMetadata;
-    @NonNull SiardContent siardContent;
+    @NonNull Metadata siardMetadata;
+    @NonNull Content siardContent;
 
     @Override
     public SiardArchive applyUpdates(Updater updater) {
