--- src/integrationTest/java/ch/admin/bar/siard2/cmd/utils/siard/model/SiardMetadata.java
@@ -1,30 +1,0 @@
-package ch.admin.bar.siard2.cmd.utils.siard.model;
-
-import ch.admin.bar.siard2.cmd.utils.siard.update.Updater;
-import ch.admin.bar.siard2.cmd.utils.siard.update.Updatable;
-import lombok.Builder;
-import lombok.Value;
-import lombok.extern.jackson.Jacksonized;
-import lombok.val;
-
-import java.util.Set;
-import java.util.stream.Collectors;
-
-@Value
-@Builder(toBuilder = true)
-@Jacksonized
-public class SiardMetadata implements Updatable<SiardMetadata> {
-    StringWrapper dbname;
-    Set<SchemaMetaData> schemas;
-
-    @Override
-    public SiardMetadata applyUpdates(Updater updater) {
-        val updatedThis = updater.applyUpdate(this);
-
-        return new SiardMetadata(
-                updatedThis.dbname.applyUpdates(updater),
-                updatedThis.schemas.stream()
-                        .map(schema -> schema.applyUpdates(updater))
-                        .collect(Collectors.toSet()));
-    }
-}
