+++ src/integrationTest/java/ch/admin/bar/siard2/cmd/utils/siard/model/header/MetadataReaderTest.java
@@ -1,0 +1,40 @@
+package ch.admin.bar.siard2.cmd.utils.siard.model.header;
+
+import ch.admin.bar.siard2.cmd.utils.SiardProjectExamples;
+import ch.admin.bar.siard2.cmd.utils.TestResourcesResolver;
+import ch.admin.bar.siard2.cmd.utils.siard.utils.Unzipper;
+import lombok.val;
+import org.assertj.core.api.Assertions;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
+
+import java.io.IOException;
+
+public class MetadataReaderTest {
+
+    @Rule
+    public TemporaryFolder temporaryFolder = new TemporaryFolder();
+
+    @Test
+    public void read_expectNoExceptions() throws IOException {
+        // given
+        val unziper = new Unzipper(
+                TestResourcesResolver.resolve(SiardProjectExamples.SAMPLE_DATALINK_2_2_SIARD),
+                temporaryFolder.getRoot());
+
+        val metadataReader = new MetadataReader(unziper.unzip());
+
+        // when
+        val result = metadataReader.read();
+
+        // then
+        Assertions.assertThat(result).isNotNull();
+
+        Assertions.assertThat(result.getSchemas()).isNotEmpty();
+        result.getSchemas().forEach(schema -> {
+            Assertions.assertThat(schema.getTables()).isNotEmpty();
+            schema.getTables().forEach(table -> Assertions.assertThat(table.getColumns()).isNotEmpty());
+        });
+    }
+}
