+++ src/integrationTest/java/ch/admin/bar/siard2/cmd/utils/siard/model/utils/Id.java
@@ -1,30 +1,31 @@
-package ch.admin.bar.siard2.cmd.utils.siard.model;
+package ch.admin.bar.siard2.cmd.utils.siard.model.utils;
 
-import ch.admin.bar.siard2.cmd.utils.siard.update.Updatable;
 import ch.admin.bar.siard2.cmd.utils.siard.update.Updater;
+import ch.admin.bar.siard2.cmd.utils.siard.update.Updatable;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import lombok.Value;
 import lombok.val;
 
 @Value(staticConstructor = "of")
-public class FolderId implements Updatable<FolderId> {
+public class Id<T> implements Updatable<Id<T>> {
     @JsonValue
-    String value;
+    StringWrapper value;
 
     @JsonCreator(mode = JsonCreator.Mode.DELEGATING)
-    public FolderId(String value) {
+    public Id(StringWrapper value) {
         this.value = value;
     }
 
     @Override
     public String toString() {
-        return value;
+        return value.toString();
     }
 
     @Override
-    public FolderId applyUpdates(Updater updater) {
+    public Id<T> applyUpdates(Updater updater) {
         val updatedThis = updater.applyUpdate(this);
-        return updatedThis;
+
+        return Id.of(updatedThis.value.applyUpdates(updater));
     }
 }
